backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.0
openai>=1.12.0
chromadb>=0.4.22
python-dotenv>=1.0.0
httpx>=0.26.0
requests>=2.31.0
wikipedia-api>=0.6.0
tiktoken>=0.6.0
